+++ design-system/storybook-addon/src/manager.tsx
@@ -1,4 +1,3 @@
-// @ts-ignore: [PIT-1685] Fails in post-office due to backwards incompatibility issue with React 18
 import { addons, types } from '@storybook/manager-api';
 
 import { ADDON_ID, TITLE, TOOL_ID } from './constants';
@@ -8,8 +7,7 @@ addons.register(ADDON_ID, () => {
 	addons.add(TOOL_ID, {
 		type: types.TOOL,
 		title: TITLE,
-		// @ts-ignore: [PIT-1685] Fails in post-office due to backwards incompatibility issue with React 18
-		match: ({ viewMode }: { viewMode: string | undefined }) =>
+		match: ({ viewMode }: { viewMode?: string | undefined }) =>
 			!!(viewMode && viewMode.match(/^(story|docs)$/)),
 		render: Tool,
 	});