+++ elements/emoji/src/components/common/Emoji.tsx
@@ -14,7 +14,6 @@ import React, {
 // eslint-disable-next-line @atlaskit/ui-styling-standard/use-compiled -- Ignored via go/DSP-18766
 import { jsx } from '@emotion/react';
 import Tooltip from '@atlaskit/tooltip';
-import { fg } from '@atlaskit/platform-feature-flags';
 import { shouldUseAltRepresentation } from '../../api/EmojiUtils';
 import {
 	deleteEmojiLabel,
@@ -166,9 +165,6 @@ const handleMouseDown = (props: Props, event: MouseEvent<any>) => {
 		return;
 	}
 	const { emoji, onSelected } = props;
-	if (!fg('editor-react-18-fix-emoji-selection')) {
-		event.preventDefault();
-	}
 	if (onSelected && leftClick(event)) {
 		onSelected(toEmojiId(emoji), emoji, event);
 	}