+++ confluence/legacy-mobile-macros/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/legacy-mobile-macros
 
+## 2.10.13
+
+### Patch Changes
+
+- Updated dependencies
+
+## 2.10.12
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 2.10.11
 
 ### Patch Changes