+++ editor/editor-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/editor-core
 
+## 195.1.6
+
+### Patch Changes
+
+- [#126974](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/pull-requests/126974)
+  [`8730c454c41f7`](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/commits/8730c454c41f7) -
+  [ux] [ED-24386] Clean up FF platform.editor.review-text-highlighting-styling
+
 ## 195.1.0
 
 ### Minor Changes